+++ src/Scope.ts
@@ -6,7 +6,7 @@ import { DiagnosticMessages } from './DiagnosticMessages';
 import type { CallableContainer, BsDiagnostic, FileReference, BscFile, CallableContainerMap, FileLink, FunctionCall, InheritableStatement, InheritableType, NamedTypeStatement } from './interfaces';
 import type { Program } from './Program';
 import { BsClassValidator } from './validators/ClassValidator';
-import type { NamespaceStatement, Statement, FunctionStatement, ClassStatement, InterfaceStatement, EnumStatement } from './parser/Statement';
+import type { NamespaceStatement, Statement, FunctionStatement, ClassStatement, EnumStatement, InterfaceStatement } from './parser/Statement';
 import type { FunctionExpression, NewExpression } from './parser/Expression';
 import { ParseMode } from './parser/Parser';
 import { standardizePath as s, util } from './util';
@@ -99,6 +99,15 @@ export class Scope {
     }
 
     /**
+     * Get the enum with the specified name.
+     * @param enumName - The enum name, including the namespace if possible
+     * @param containingNamespace - The namespace used to resolve relative enum names. (i.e. the namespace around the current statement trying to find an enum)
+     */
+    public getEnum(enumName: string, containingNamespace?: string): EnumStatement {
+        return this.getEnumFileLink(enumName, containingNamespace)?.item;
+    }
+
+    /**
      * Get a class and its containing file by the class name
      * @param className - The class name, including the namespace of the class if possible
      * @param containingNamespace - The namespace used to resolve relative class names. (i.e. the namespace around the current statement trying to find a class)
@@ -116,6 +125,7 @@ export class Scope {
         }
         return cls;
     }
+
 
     /**
    * Get an interface and its containing file by the interface name
@@ -228,7 +238,7 @@ export class Scope {
 
     /**
     * Tests if an interface exists with the specified name
-    * @param ifaceName - the all-lower-case namespace-included class name
+    * @param ifaceName - the all-lower-case namespace-included interface name
     * @param namespaceName - the current namespace name
     */
     public hasInterface(ifaceName: string, namespaceName?: string): boolean {
@@ -236,10 +246,19 @@ export class Scope {
     }
 
     /**
-    * Tests if a class OR an interface, etc. exists with the specified name
-    * @param name - the all-lower-case namespace-included class or interface name
+    * Tests if an enum exists with the specified name
+    * @param enumName - the all-lower-case namespace-included enum name
     * @param namespaceName - the current namespace name
     */
+    public hasEnum(enumName: string, namespaceName?: string): boolean {
+        return !!this.getEnum(enumName, namespaceName);
+    }
+
+    /**
+     * Tests if a class OR an interface, etc. exists with the specified name
+     * @param name - the all-lower-case namespace-included class or interface name
+     * @param namespaceName - the current namespace name
+     */
     public hasNamedType(name: string, namespaceName?: string): boolean {
         return !!this.getNamedTypeStatement(name, namespaceName);
     }
@@ -275,19 +294,19 @@ export class Scope {
         return [];
     }
     /**
-   * A dictionary of all Interfaces in this scope. This includes namespaced Interfaces always with their full name.
-   * The key is stored in lower case
-   */
+    * A dictionary of all Interfaces in this scope. This includes namespaced Interfaces always with their full name.
+    * The key is stored in lower case
+    */
     public getInterfaceMap(): Map<string, FileLink<InterfaceStatement>> {
         return this.cache.getOrAdd('interfaceMap', () => {
             const map = new Map<string, FileLink<InterfaceStatement>>();
             this.enumerateBrsFiles((file) => {
                 if (isBrsFile(file)) {
-                    for (let cls of file.parser.references.interfaceStatements) {
-                        const lowerClassName = cls.getName(ParseMode.BrighterScript)?.toLowerCase();
+                    for (let iface of file.parser.references.interfaceStatements) {
+                        const lowerIfaceName = iface.getName(ParseMode.BrighterScript)?.toLowerCase();
                         //only track classes with a defined name (i.e. exclude nameless malformed classes)
-                        if (lowerClassName) {
-                            map.set(lowerClassName, { item: cls, file: file });
+                        if (lowerIfaceName) {
+                            map.set(lowerIfaceName, { item: iface, file: file });
                         }
                     }
                 }
@@ -916,7 +935,7 @@ export class Scope {
                 if (!paramType && param.type) {
                     const paramTypeName = param.type.getText();
                     const currentNamespaceName = func.namespaceName?.getName(ParseMode.BrighterScript);
-                    if (!this.hasClass(paramTypeName, currentNamespaceName)) {
+                    if (!this.hasClass(paramTypeName, currentNamespaceName) && !this.hasInterface(paramTypeName) && !this.hasEnum(paramTypeName)) {
                         this.diagnostics.push({
                             ...DiagnosticMessages.functionParameterTypeIsInvalid(param.name.text, paramTypeName),
                             range: param.type.range,